Add cache for deployed pieces.
When calculating all possible moves many positions will be recalculated
after a piece has been deployed. To reduce the overhead every piece in
all it's variations is stored in 'allPieces'. Therefore the move
calculation can look the piece up and return a pointer of the actual
piece. This method prevents recalculation and memory copies.

The performance increase is about 82%.