Add a cache to canBeDeployed
Blokus is a game where pieces can be deployed but not moved. Therefore
a piece which cannot be deployed at state x cannot be deployed at state
x + 1 either. This allows to cache if a piece cannot be deployed.
If a piece is not explicit marked as undeployable it must be checked
whether it can be deployed or not.

When reverting a move some undeployable pieces can be deployed again.
That is why the game state needs to store a history of undeployable
positions.

The performance increase is about 25%.