Filter duplicated moves + Improve debug print
In some game situations multiple attach points of the same piece can
lead to the same deployed piece. To filter those moves out of the
dataset improves the overall search performance because less (identical)
branches have to be searched.

The performance of the move calculation is improved too. This is an
effect of less expensive accesses to `allPieces`.