Fix: Piece rotations conform to official rules
In previous versions the piece shapes where fliped first and afterwards
rotated. According to the official rules they have to be rotated before
they are fliped. If done otherwise the piece shapes are in a different
order than expected by the game server.

Example: Client sends a valid move with rotation PI_FLIPPED. The server
interprets it as THREEHALFPI_FLIPPED. Therefore the move is invalid.