Change how the current color is determined
Using the current turn as indication of the current color will lead to
wrong colors and therefore disqualification. This happens because the
turn sent by the server only increments when a client sends a move, not
when the server skips the move of a client by itself. Fortunately the
memnto message contains the valid index of the current color. The index
can be used to retrieve the color from `colorInGame` also sent by the
server